+++ BambooTracker/command/pattern/set_key_on_to_step_command.cpp
@@ -1,7 +1,7 @@
 #include "set_key_on_to_step_command.hpp"
 
 SetKeyOnToStepCommand::SetKeyOnToStepCommand(std::weak_ptr<Module> mod, int songNum, int trackNum, int orderNum,
-											 int stepNum, int noteNum, bool autosetInst, int instNum)
+											 int stepNum, int noteNum, bool instMask, int instNum, bool volMask, int vol)
 	: mod_(mod),
 	  song_(songNum),
 	  track_(trackNum),
@@ -9,25 +9,30 @@ SetKeyOnToStepCommand::SetKeyOnToStepCommand(std::weak_ptr<Module> mod, int song
 	  step_(stepNum),
 	  note_(noteNum),
 	  inst_(instNum),
-	  autosetInst_(autosetInst)
+	  vol_(vol),
+	  instMask_(instMask),
+	  volMask_(volMask)
 {
 	auto& st = mod_.lock()->getSong(songNum).getTrack(trackNum).getPatternFromOrderNumber(orderNum).getStep(stepNum);
 	prevNote_ = st.getNoteNumber();
-	if (autosetInst) prevInst_ = st.getInstrumentNumber();
+	if (!instMask) prevInst_ = st.getInstrumentNumber();
+	if (!volMask) prevVol_ = st.getVolume();
 }
 
 void SetKeyOnToStepCommand::redo()
 {
 	auto& st = mod_.lock()->getSong(song_).getTrack(track_).getPatternFromOrderNumber(order_).getStep(step_);
 	st.setNoteNumber(note_);
-	if (autosetInst_) st.setInstrumentNumber(inst_);
+	if (!instMask_) st.setInstrumentNumber(inst_);
+	if (!volMask_) st.setVolume(vol_);
 }
 
 void SetKeyOnToStepCommand::undo()
 {
 	auto& st = mod_.lock()->getSong(song_).getTrack(track_).getPatternFromOrderNumber(order_).getStep(step_);
 	st.setNoteNumber(prevNote_);
-	if (autosetInst_) st.setInstrumentNumber(prevInst_);
+	if (!instMask_) st.setInstrumentNumber(prevInst_);
+	if (!volMask_) st.setVolume(prevVol_);
 }
 
 CommandId SetKeyOnToStepCommand::getID() const