+++ BambooTracker/command/pattern/set_key_on_to_step_command.hpp
@@ -8,14 +8,14 @@ class SetKeyOnToStepCommand : public AbstractCommand
 {
 public:
 	SetKeyOnToStepCommand(std::weak_ptr<Module> mod, int songNum, int trackNum, int orderNum, int stepNum,
-						  int noteNum, bool autosetInst, int instNum);
+						  int noteNum, bool instMask, int instNum, bool volMask, int vol);
 	void redo() override;
 	void undo() override;
 	CommandId getID() const override;
 
 private:
 	std::weak_ptr<Module> mod_;
-	int song_, track_, order_, step_, note_, inst_;
-	int prevNote_, prevInst_;
-	bool autosetInst_;
+	int song_, track_, order_, step_, note_, inst_, vol_;
+	int prevNote_, prevInst_, prevVol_;
+	bool instMask_, volMask_;
 };